Fix bias amount, add cubemap texture

+++ src/RendererEngine/Components/RenderBuffer.cpp
@@ -30,7 +30,7 @@ namespace RenderingEngine
     {
         glCreateBuffers(1, &m_RendererID);
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_RendererID);
-        glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(int), indices, GL_STATIC_DRAW);
+        glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(float), indices, GL_STATIC_DRAW);
 
         m_IndexCount = count;
     }
